@@ -3,19 +3,23 @@ import { BrowserModule } from '@angular/platform-browser';
 import { RouteReuseStrategy } from '@angular/router';
 
 import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
-import { HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { FileOpener } from '@ionic-native/file-opener/ngx';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
-import { CurrencyPipe, DatePipe, registerLocaleData } from '@angular/common';
+import { AsyncPipe, CommonModule, CurrencyPipe, DatePipe, registerLocaleData } from '@angular/common';
 import localDe from '@angular/common/locales/de';
 import { ReactiveFormsModule } from '@angular/forms';
+import { AuthHttpInterceptorService } from './auth/auth-http-interceptor.service';
+import { SettlementPaymentPage } from './settlement-payment/settlement-payment.page';
+import { GroupMembersPage } from './group-members/group-members.page';
 registerLocaleData(localDe, 'de');
 
 @NgModule({
-  declarations: [AppComponent],
+  declarations: [AppComponent, SettlementPaymentPage, GroupMembersPage],
   imports: [
     BrowserModule,
+    CommonModule,
     IonicModule.forRoot(), 
     AppRoutingModule,
     HttpClientModule,
@@ -24,7 +28,9 @@ registerLocaleData(localDe, 'de');
   providers: [
     DatePipe,
     CurrencyPipe,
+    AsyncPipe,
     FileOpener,
+    { provide: HTTP_INTERCEPTORS, useClass: AuthHttpInterceptorService, multi: true },
     { provide: RouteReuseStrategy, useClass: IonicRouteStrategy }
   ],
   bootstrap: [AppComponent],
